video: tegra: host: Fix build warnings
Add -Wno-multichar to the directories where we use multichar constants
in a legitimite way.

Increase cyclestat_buffer_size parameter to u64.

Remove some unused variables and commented code.

Fix function signatures for gk20a clock code.

Convert Makefiles to use ccflags-y instead of EXTRA_CFLAGS.

Reviewed-on: http://git-master/r/263240

+++ drivers/video/tegra/host/gk20a/channel_gk20a.c
@@ -423,7 +423,7 @@ int gk20a_channel_cycle_stats(struct channel_gk20a *ch,
 	struct mem_mgr *memmgr = gk20a_channel_mem_mgr(ch);
 	struct mem_handle *handle_ref;
 	void *virtual_address;
-	u32 cyclestate_buffer_size;
+	u64 cyclestate_buffer_size;
 	struct platform_device *dev = ch->ch->dev;
 
 	if (args->nvmap_handle && !ch->cyclestate.cyclestate_buffer_handler) {
